[CORE] Backport PIVX changes.

+++ src/logging.h
@@ -30,7 +30,7 @@ struct CLogCategoryActive
 };
 
 namespace BCLog {
-    enum LogFlags : uint32_t {
+    enum LogFlags : uint64_t {
         NONE        = 0,
         NET         = (1 <<  0),
         TOR         = (1 <<  1),
@@ -62,9 +62,9 @@ namespace BCLog {
         ZWSP        = (1 << 27),
         ALERT       = (1 << 28),
         POS         = (1 << 29),
-        SPORK         = (1 << 30),
-        PRECOMPUTE         = (1 << 31),
-        ALL         = ~(uint32_t)0,
+        SPORK       = (1 << 30),
+        PRECOMPUTE  = (1 << 31),
+        ALL         = ~(uint64_t)0,
     };
 
     class Logger
